Added more weapons: Swarm missile which launches multiple rockets roughly in direction of mouse that start with low speed and accelerate. bluemerange which is a blue bomerange of which there are tue.
Rotation speed is now a float. (used to be int)

Changed the bee holding sprite

Both weapons added to tree.

+++ source/core/src/main/com/csse3200/game/entities/configs/WeaponConfig.java
@@ -13,7 +13,7 @@ public class WeaponConfig extends BaseEntityConfig {
     //Weapon stats
     public float damage = 0;
     public float weaponSpeed = 0;
-    public int rotationSpeed = 0;
+    public float rotationSpeed = 0;
     public int weaponDuration = 0;
     public int attackCooldown = 5;
     public int ammoUse = 1;
